Move OrderIntent enum out of SDK and into Demo app.
