+++ Demo/src/main/java/com/paypal/android/api/services/SDKSampleServerAPI.kt
@@ -5,7 +5,7 @@ import com.google.gson.JsonParser
 import com.paypal.android.api.model.ClientId
 import com.paypal.android.api.model.CreateOrderRequest
 import com.paypal.android.api.model.Order
-import com.paypal.android.cardpayments.OrderIntent
+import com.paypal.android.api.model.OrderIntent
 import com.paypal.android.usecase.UpdateOrderUseCase
 import com.paypal.checkout.order.OrderRequest
 import okhttp3.OkHttpClient
