+++ Demo/src/main/java/com/paypal/android/ui/card/CardViewModel.kt
@@ -2,7 +2,7 @@ package com.paypal.android.ui.card
 
 import androidx.lifecycle.ViewModel
 import com.paypal.android.api.model.Order
-import com.paypal.android.cardpayments.OrderIntent
+import com.paypal.android.api.model.OrderIntent
 import com.paypal.android.cardpayments.model.CardResult
 import com.paypal.android.models.TestCard
 import com.paypal.android.ui.card.validation.CardViewUiState
