+++ Demo/src/main/java/com/paypal/android/ui/card/validation/CardViewUiState.kt
@@ -2,7 +2,7 @@ package com.paypal.android.ui.card.validation
 
 import androidx.compose.runtime.Immutable
 import com.paypal.android.api.model.Order
-import com.paypal.android.cardpayments.OrderIntent
+import com.paypal.android.api.model.OrderIntent
 import com.paypal.android.cardpayments.model.CardResult
 
 @Immutable
