+++ Demo/src/main/java/com/paypal/android/ui/createorder/CreateOrderUiState.kt
@@ -1,6 +1,6 @@
 package com.paypal.android.ui.createorder
 
-import com.paypal.android.cardpayments.OrderIntent
+import com.paypal.android.api.model.OrderIntent
 
 data class CreateOrderUiState(
     val intentOption: OrderIntent = OrderIntent.AUTHORIZE,
